+++ bindings/python/examples/client/07_get_block_data.py
@@ -1,3 +1,4 @@
+from dataclasses import asdict
 from iota_sdk import Client
 from dotenv import load_dotenv
 import json
@@ -16,8 +17,8 @@ print(f'Block id: {block_ids[0]}')
 
 # Get the metadata for the block
 metadata = client.get_block_metadata(block_ids[0])
-print(f'Block metadata: {json.dumps(metadata, indent=4)}')
+print(f'Block metadata: {json.dumps(asdict(metadata), indent=4)}')
 
 # Request the block by its id
 block = client.get_block_data(block_ids[0])
-print(f'Block: {json.dumps(block, indent=4)}')
+print(f'Block: {json.dumps(asdict(block), indent=4)}')