+++ bindings/python/examples/client/custom_plugin.py
@@ -25,5 +25,6 @@ nft_id = "0x0000000000000000000000000000000000000000000000000000000000000000"
 route = "outputs/nft/" + nft_id
 
 # Call our "custom" indexer plugin
-output_id = client.call_plugin_route("api/indexer/v1/", "GET", route, None, None)
+output_id = client.call_plugin_route(
+    "api/indexer/v1/", "GET", route, None, None)
 print(json.dumps(output_id, indent=4))