+++ bindings/python/examples/wallet/migrate-stronghold-snapshot-v2-to-v3.py
@@ -1,5 +1,5 @@
 import iota_sdk
-from iota_sdk import Wallet, StrongholdSecretManager, CoinType
+from iota_sdk import Wallet, StrongholdSecretManager, CoinType, ClientOptions
 from dotenv import load_dotenv
 import os
 
@@ -8,9 +8,7 @@ load_dotenv()
 v2_path = "../../../sdk/tests/wallet/fixtures/v2.stronghold"
 v3_path = "./v3.stronghold"
 node_url = os.environ.get('NODE_URL', 'https://api.testnet.shimmer.network')
-client_options = {
-    'nodes': [node_url],
-}
+client_options = ClientOptions(nodes=[node_url])
 coin_type = CoinType.SHIMMER
 
 try: